backend/requirements.txt
fastapi==0.115.12
google==3.0.0
google-api-core==2.24.2
google-genai==1.12.1
grpcio==1.71.0
pinecone==6.0.2
protobuf==6.30.2
python-dotenv==1.1.0
uvicorn==0.34.2
